build.mk script to use kos-cmake wrapper.
The build system for kos-ports currently raw-dogs "cmake" as the command
to build a CMakeLists.txt file, as opposed to our preferred "kos-cmake"
wrapper file.

In addition to this requiring us to manually pass the toolchain argument
for each invocation, it also means that should kos-cmake need to change
in the future, the changes would not permeate out.

I just made a bugfix for DreamSDK which forces cmake to use the "Unix
Makefiles" generator by explicitly passing it to cmake as part of the
kos-cmake wrapper. This is required to ensure that CMake's top generator
choice on Windows, freaking Visual Studio, does not get used when trying
to build kos-ports.