Added libSH4ZAM to kos-ports.
Even though we want to eventually make SH4ZAM be a kos add-on, which
will allow us to more tightly couple it with libkallisti.a, there's
still a lot of design and implementation work up in the air for figuring
out how the new model should work that is gating my ability to introduce
SH4ZAM into the greater KOS ecosystem.

At this point, with its current userbase, maturity, and proven gainz, I
felt like it was in everyone's best interest to get it introduced as a
kos-port initially, using the old model, so that people can already be
using it and other kos-ports can already be depending on it for
acceleration without us having to rush to implement the new KOS add-on
model... Which I will happily migrate to as soon as it's ready, and I
believe that moving SH4ZAM from a "kos-port" to a kos-addon from a user
perspective should be transparent and non-breaking in the future.

ANYWAY:
1) Added pkg-descr with high-level README information on what SH4ZAM is
   and does for readers and for DreamSDK's UI, along with URL for
   the project documentation.
2) Added Makefile which builds the library using the CMake build system
   (similarly to what we're doing for GLdc), then installs the libs and
   public API headers, which I have confirmed is done properly.
3) "Bruce's Balls" example should also get installed to
   "kos-ports/examples/sh4zam/bruces_balls" which serves as an example,
   benchmark, and real-world SH4ZAM test... also its' a kick-ass PVR DR
   high-performance rendering example independently of SH4ZAM.